backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0

pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart

pillow==10.4.0
opencv-python
numpy

torch==2.6.0
torchvision==0.21.0
ultralytics

pytesseract

sqlalchemy
pymysql
